Migrate from SQLite3 to PostgreSQL
It is a best practice to have the development database be the same as the (intended) production database. And it is easier to switch at this early point, before I have any data in the database. The ultimate steps that were necessary for the migration:

* `rails new CAC-Student-Podcasts -d postgresql` in a different folder
* Change from the 'sqlite3' gem to the 'pg' gem in this app
* Copy the config/database.yml from the new app. (No need to find and replace the app name, since the temporary new project has the same name.)
* Run `rake db:setup` to set up the development database
* delete the old db/*.sqlite3 development database files

On my Mac, to make local development almost as convenient as with SQLite, I installed Postgres.app to launch the server, and PG Commander (like Sequel Pro) to view tables.